fix: move dock clicker too other files
